python 3.11.4 macos

+++ python/3.11.4/macos/install.sh
@@ -5,7 +5,10 @@ PKG_CONFIG_FOLDER_NAME=pkg-config
 PKG_CONFIG_VERSION=0.29.2
 
 GETTEXT_FOLDER_NAME=gettext
-GETTEXT_VERSION=0.22
+GETTEXT_VERSION=0.20.1
+
+OPENSSL_VERSION=3.1.2
+OPENSSL_FOLDER_NAME=openssl
 
 INSTALL_FILES_DIR=$HOME/install-files
 
@@ -17,6 +20,10 @@ if [ ! -d "$HOME/programs" ]; then
 	mkdir "$HOME/programs"
 fi
 
+if [ ! -d "$HOME/logs" ]; then
+	mkdir "$HOME/logs"
+fi
+
 if [ ! -d "$HOME/sources/$FOLDER_NAME" ]; then
 	mkdir "$HOME/sources/$FOLDER_NAME"
 fi
@@ -25,11 +32,20 @@ if [ ! -d "$HOME/programs/$FOLDER_NAME" ]; then
 	mkdir "$HOME/programs/$FOLDER_NAME"
 fi
 
+if [ ! -d "$HOME/logs/$FOLDER_NAME" ]; then
+	mkdir "$HOME/logs/$FOLDER_NAME"
+fi
+
+if [ ! -d "$HOME/logs/$FOLDER_NAME/$VERSION" ]; then
+	mkdir "$HOME/logs/$FOLDER_NAME/$VERSION"
+fi
+
 if [ ! -d "$HOME/programs/$FOLDER_NAME/$VERSION" ]; then
 	mkdir "$HOME/programs/$FOLDER_NAME/$VERSION"
 
 	bash $INSTALL_FILES_DIR/$PKG_CONFIG_FOLDER_NAME/$PKG_CONFIG_VERSION/macos/install.sh
 	bash $INSTALL_FILES_DIR/$GETTEXT_FOLDER_NAME/$GETTEXT_VERSION/macos/install.sh
+	bash $INSTALL_FILES_DIR/$OPENSSL_FOLDER_NAME/$OPENSSL_VERSION/macos/install.sh
 
 	cd $HOME/sources/$FOLDER_NAME
 
@@ -38,18 +54,29 @@ if [ ! -d "$HOME/programs/$FOLDER_NAME/$VERSION" ]; then
 	export LDFLAGS="-L$HOME/programs/$GETTEXT_FOLDER_NAME/$GETTEXT_VERSION/lib"
 	export CPPFLAGS="-I$HOME/programs/$GETTEXT_FOLDER_NAME/$GETTEXT_VERSION/include"
 
-	wget "https://www.python.org/ftp/python/"$VERSION"/Python-"$VERSION".tgz"
-	tar -xvf "Python-"$VERSION".tgz"
+	printf "${bold}${yellow}Installing $FOLDER_NAME${clear}\n"
+
+	printf "\t${bold}${green}Downloading source code${clear}\n"
+	wget -q --show-progress "https://www.python.org/ftp/python/"$VERSION"/Python-"$VERSION".tgz"
+	printf "\t${bold}${green}Extracting source code${clear}\n"
+	tar -xf "Python-"$VERSION".tgz"
 	mv "Python-"$VERSION $VERSION
 	cd $VERSION
-	./configure --with-pydebug --prefix="$HOME/programs/python/$VERSION"
-	make -s -j2
-	sudo make install
+	printf "\t${bold}${green}Configuring${clear}\n"
+	./configure --help > $HOME/logs/$FOLDER_NAME/$VERSION/configureHelp.txt 2>&1
+	./configure --with-pydebug --prefix="$HOME/programs/python/$VERSION" --with-openssl=$HOME/programs/$OPENSSL_FOLDER_NAME/$OPENSSL_VERSION > $HOME/logs/$FOLDER_NAME/$VERSION/configureOutput.txt 2>&1
+	printf "\t${bold}${green}Making${clear}\n"
+	make -s -j2 > $HOME/logs/$FOLDER_NAME/$VERSION/makeOutput.txt 2>&1
+	printf "\t${bold}${green}Installing${clear}\n"
+	echo $USER_PASSWORD | sudo -S -p '' make install > $HOME/logs/$FOLDER_NAME/$VERSION/installOutput.txt 2>&1
 
-	cd $HOME/programs/$FOLDER_NAME/$VERSION
-	sudo chown -R $(whoami) .
+	if [ -e "$HOME/programs/$FOLDER_NAME/$VERSION/bin/python3" ]; then
+		cd $HOME/programs/$FOLDER_NAME/$VERSION
+		echo $USER_PASSWORD | sudo -S -p '' chown -R $(whoami) .
 
-	cd $HOME/sources/$FOLDER_NAME
-	sudo rm -rf $VERSION
-	rm "Python-"$VERSION".tgz"
+		printf "\t${bold}${green}Clearing${clear}\n"
+		cd $HOME/sources/$FOLDER_NAME
+		sudo rm -rf $VERSION
+		rm "Python-"$VERSION".tgz"
+	fi
 fi
